Extract notifications settings into own view and add tasks panel to Time Clock
- Move notification settings (clock-in/out reminders, time pickers, scheduling
  logic) from MoreView into new NotificationsSettingsView, accessed via
  NavigationLink
- Add tasks editing panel to TimeClockView that appears during active sessions,
  showing current tasks with project names and an "Edit Tasks" button that
  opens the TasksEditorView sheet
- Add localized strings for new UI elements (TimeClock.Tasks.Title,
  TimeClock.Tasks.Edit, Settings.Notifications.ClockIn/ClockOut) in both
  English and Japanese
