+++ WorkingHour/Views/More/MoreView.swift
@@ -20,21 +20,6 @@ struct MoreView: View {
     @State private var breakMinutes: Double = 60.0
     @State private var autoAddBreak: Bool = false
 
-    @State private var clockInReminderEnabled: Bool = false
-    @State private var clockOutReminderEnabled: Bool = false
-    @State private var clockInReminderTime: Date = {
-        var components = DateComponents()
-        components.hour = 8
-        components.minute = 0
-        return Calendar.current.date(from: components) ?? Date()
-    }()
-    @State private var clockOutReminderTime: Date = {
-        var components = DateComponents()
-        components.hour = 17
-        components.minute = 0
-        return Calendar.current.date(from: components) ?? Date()
-    }()
-
     var body: some View {
         NavigationStack {
             MoreList(repoName: "katagaki/WorkingHour") {
@@ -107,18 +92,8 @@ struct MoreView: View {
 
                 // Notifications Section
                 Section {
-                    Toggle("Settings.ClockInReminder", isOn: $clockInReminderEnabled)
-                    if clockInReminderEnabled {
-                        DatePicker("Settings.ClockInReminder.Time",
-                                   selection: $clockInReminderTime,
-                                   displayedComponents: .hourAndMinute)
-                    }
-
-                    Toggle("Settings.ClockOutReminder", isOn: $clockOutReminderEnabled)
-                    if clockOutReminderEnabled {
-                        DatePicker("Settings.ClockOutReminder.Time",
-                                   selection: $clockOutReminderTime,
-                                   displayedComponents: .hourAndMinute)
+                    NavigationLink("Settings.Section.Notifications") {
+                        NotificationsSettingsView()
                     }
                 } header: {
                     Text("Settings.Section.Notifications")
@@ -191,18 +166,6 @@ struct MoreView: View {
             .onChange(of: autoAddBreak) { _, _ in
                 saveSettings()
             }
-            .onChange(of: clockInReminderEnabled) { _, _ in
-                saveNotificationSettings()
-            }
-            .onChange(of: clockOutReminderEnabled) { _, _ in
-                saveNotificationSettings()
-            }
-            .onChange(of: clockInReminderTime) { _, _ in
-                saveNotificationSettings()
-            }
-            .onChange(of: clockOutReminderTime) { _, _ in
-                saveNotificationSettings()
-            }
         }
     }
 
@@ -210,59 +173,11 @@ struct MoreView: View {
         workingHours = settingsManager.standardWorkingHours / 3600.0
         breakMinutes = settingsManager.defaultBreakDuration / 60.0
         autoAddBreak = settingsManager.autoAddBreakTime
-        clockInReminderEnabled = settingsManager.clockInReminderEnabled
-        clockOutReminderEnabled = settingsManager.clockOutReminderEnabled
-        clockInReminderTime = dateFromSecondsSinceMidnight(settingsManager.clockInReminderTime)
-        clockOutReminderTime = dateFromSecondsSinceMidnight(settingsManager.clockOutReminderTime)
     }
 
     private func saveSettings() {
         settingsManager.standardWorkingHours = workingHours * 3600
         settingsManager.defaultBreakDuration = breakMinutes * 60
         settingsManager.autoAddBreakTime = autoAddBreak
-    }
-
-    private func saveNotificationSettings() {
-        settingsManager.clockInReminderEnabled = clockInReminderEnabled
-        settingsManager.clockOutReminderEnabled = clockOutReminderEnabled
-        settingsManager.clockInReminderTime = secondsSinceMidnight(from: clockInReminderTime)
-        settingsManager.clockOutReminderTime = secondsSinceMidnight(from: clockOutReminderTime)
-
-        let notificationManager = NotificationManager.shared
-
-        if clockInReminderEnabled {
-            let components = Calendar.current.dateComponents([.hour, .minute], from: clockInReminderTime)
-            Task {
-                await notificationManager.scheduleClockInReminders(at: components)
-                settingsManager.notificationsLastScheduledDate = Date()
-            }
-        } else {
-            Task {
-                await notificationManager.cancelClockInReminders()
-            }
-        }
-
-        if clockOutReminderEnabled {
-            let components = Calendar.current.dateComponents([.hour, .minute], from: clockOutReminderTime)
-            Task {
-                await notificationManager.scheduleClockOutReminder(at: components)
-            }
-        } else {
-            notificationManager.cancelClockOutReminder()
-        }
-    }
-
-    private func secondsSinceMidnight(from date: Date) -> Double {
-        let components = Calendar.current.dateComponents([.hour, .minute], from: date)
-        return Double((components.hour ?? 0) * 3600 + (components.minute ?? 0) * 60)
-    }
-
-    private func dateFromSecondsSinceMidnight(_ seconds: Double) -> Date {
-        let hour = Int(seconds) / 3600
-        let minute = (Int(seconds) % 3600) / 60
-        var components = DateComponents()
-        components.hour = hour
-        components.minute = minute
-        return Calendar.current.date(from: components) ?? Date()
     }
 }
